simplify the parsing pipeline, update readme
