+++ package.json
@@ -5,6 +5,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
+    "publish": "npx rimraf dist && npm run build && npm version patch && npm publish --access public",
     "build": "tsc",
     "test": "jest"
   },
