+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './parser';
-export * from './resolver';
 export * from './utils';
