+++ src/parser/RichTextBaseParser.ts
@@ -1,18 +1,15 @@
-import { IParser, IParseResult, IParserEngine } from "./parser-models";
+import { IParserEngine, RichTextInput, IOutputResult, IRichTextParser } from "./parser-models";
 import { isRootNode } from "../utils/";
 
-export abstract class RichTextBaseParser implements IParser {
+export abstract class RichTextBaseParser implements IRichTextParser<RichTextInput, IOutputResult> {
     abstract _parser: IParserEngine;
     abstract parseInternal(node: any): any;
 
-    parse(value: string): IParseResult {
-        const document = this._parser.parse(value);
+    parse(input: RichTextInput): IOutputResult {
+        const node = this._parser.parse(input.value);
 
-        if (isRootNode(document) && document.body.firstChild) {
-            return {
-                children: Array.from(document.body.childNodes).flatMap((node) => this.parseInternal(node))
-            }
-
+        if (isRootNode(node) && node.body.firstChild) {
+            throw new Error("Not implemented.")
         }
         else {
             throw new Error();
