+++ src/parser/parser-models.ts
@@ -1,9 +1,5 @@
 export type IDomNode = IDomHtmlNode | IDomTextNode;
 
-export type IParseResult = {
-    children: IDomNode[]
-}
-
 export interface IDomTextNode {
     type: 'text'
     content: string
@@ -16,12 +12,36 @@ export interface IDomHtmlNode {
     children: IDomNode[]
 }
 
-export interface IParser {
-    parse(value: string): {
-        children: IDomNode[]
-    }
-}
-
 export interface IParserEngine {
     parse(html: string): any;
+}
+
+export type RichTextInput = {
+    value: string,
+    images?: {
+        [key: string]: {
+            image_id?: string,
+            description?: string | null,
+            url: string,
+            width?: number | null, // TODO changed from undefined for easier SDK conversion. potential problems?
+            height?: number | null
+        }
+    },
+    links?: {
+        [key: string]: {
+            codename: string,
+            type?: string,
+            url_slug?: string
+        }
+    },
+    modular_content?: string[],
+    linked_items?: { [key: string]: {} }
+}
+
+export interface IOutputResult {
+    childNodes: IDomNode[]
+}
+
+export interface IRichTextParser<TInput, TOutput> {
+    parse(input: TInput): TOutput
 }
