--- src/resolver/index.ts
@@ -1,2 +1,0 @@
-export * from './RichTextResolver';
-export * from './resolver-models';
